Revert "Sample change"
