--- .github/workflows/deploy.yml
@@ -1,77 +1,0 @@
-name: Deploy
-
-on:
-  push:
-    branches:
-      - main
-  workflow_dispatch:
-  pull_request:
-    branches:
-      - main
-    types:
-      - closed
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    name: Run Tests
-    env:
-      TESTING: true
-    steps:
-      - name: Checkout Repository
-        uses: actions/checkout@v2
-
-      - name: Setup Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.8.9'
-
-      - name: Setup Python Virtual Environment
-        run: python -m venv python3-virtualenv
-
-      - name: Install Dependencies
-        run: python3-virtualenv/bin/pip install -r requirements.txt
-
-      - name: Run Tests
-        run: ./run_test.sh
-
-      - name: Notify on fail
-        if: ${{ failure() }}
-        run: curl -s -X POST "${{ secrets.DISCORD_WEBHOOK }}" -d "content=🚨 Testing failed ($GITHUB_ACTION - id=$GITHUB_RUN_ID)"
-
-
-  deploy:
-    name: "Deploy to VPS"
-    runs-on: ubuntu-latest
-    needs: test
-    steps:
-      - name: Configure SSH
-        run: |
-          mkdir -p ~/.ssh/
-          echo "$SSH_PRIVATE_KEY" > ~/.ssh/deploy-key.pem
-          chmod 600 ~/.ssh/deploy-key.pem
-          cat >> ~/.ssh/config <<END
-          Host my-vps
-            HostName $SSH_IP
-            User $SSH_USER
-            IdentityFile ~/.ssh/deploy-key.pem
-            StrictHostKeyChecking no
-          END
-        env:
-          SSH_USER: ${{ secrets.SSH_USER }}
-          SSH_PRIVATE_KEY: ${{ secrets.SSH_PRIVATE_KEY }}
-          SSH_IP: ${{ secrets.SSH_IP }}
-
-      - name: Deploy site
-        run: ssh my-vps '~/redeploy-site.sh'
-
-      - name: Print out container status
-        run: ssh my-vps 'cd project-flask-falcons && docker compose ps'
-
-      - name: Notify on success
-        if: ${{ success() }}
-        run: curl -s -X POST "${{ secrets.DISCORD_WEBHOOK }}" -d "content=🚀 Deployment Successful (id $GITHUB_RUN_ID)"
-
-      - name: Notify on fail
-        if: ${{ failure() }}
-        run: curl -s -X POST "${{ secrets.DISCORD_WEBHOOK }}" -d "content=🚨 Deployment failed (id $GITHUB_RUN_ID)"
