--- .github/workflows/test.yml
@@ -1,28 +1,0 @@
-name: Run Tests
-on:
-  pull_request:
-    branches:
-      - main
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    name: Run Tests
-    env:
-      TESTING: true
-    steps:
-      - name: Checkout Repository
-        uses: actions/checkout@v2
-
-      - name: Setup Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.8.9'
-
-      - name: Setup Python Virtual Environment
-        run: python -m venv python3-virtualenv
-
-      - name: Install Dependencies
-        run: python3-virtualenv/bin/pip install -r requirements.txt
-
-      - name: Run Tests
-        run: ./run_test.sh
