--- app/static/data/oscar_experience.json
@@ -1,16 +1,0 @@
-[
-  {
-    "title": "Production Engineering Fellow",
-    "date": "May 2002 - present",
-    "organization": "MLH",
-    "logo_url": "https://external-content.duckduckgo.com/iu/?u=https%3A%2F%2Fstatic.mlh.io%2Fbrand-assets%2Flogo%2Fofficial%2Fmlh-logo-color.png&f=1&nofb=1",
-    "description": "- Deployed application to VPS"
-  },
-  {
-    "title": "Coordinator for chapter at Tec Guadalajara Campus",
-    "date": "August 2021 - June 2022",
-    "organization": "Chamber of Commerce of Guadalajara",
-    "logo_url": "https://external-content.duckduckgo.com/iu/?u=http%3A%2F%2Fwww.licigob.com.mx%2FCGI-BIN%2Flogo_canaco.jpg&f=1&nofb=1",
-    "description": "- Managed student memberships\n- Event planning and execution\n- Process improvements"
-  }
-]
