+++ app/static/styles/main.css
@@ -5,52 +5,33 @@ p, h1, h2, h3, h4, h5, h6, button {
 }
 
 body {
-    color: white;
-    background-color: #00264b;
+    background-color: #e7ffd5;
     background-size: cover;
     width: auto;
     overflow-x: hidden;
-    font-family: "Raleway", serif;
+    font-family: "Roboto", serif;
     margin: 0px;
 }
 
-header {
-    padding: 50px 20px 20px 20px;
-    margin: 0 auto;
-    height: 40px;
-    max-width: 800px;
-
+.nav-bar {
+    max-width: 1200px;
     display: flex;
     justify-content: space-between;
+    align-items: center;
+    font-family: "Roboto", serif;
+    font-weight: 700 !important;
+    padding: 20px;
+    margin: 0 auto;
+    height: 60px;
 }
 
-.site-title-header {
+.nav-bar a {
     text-decoration: none;
-    margin-bottom: 0;
+    color: black;
 }
 
-.site-title-header:visited {
-    color: white;
-}
-
-@media (max-width: 600px) {
-    header {
-        padding: 20px;
-        height: 20px;
-    }
-}
-
-header a {
-    text-decoration: none;
-    color: white;
-}
-
-header p {
-    margin: 0;
-}
-
-header a:visited {
-    color: white;
+.nav-bar a:visited {
+    color: black;
 }
 
 
@@ -60,16 +41,10 @@ header a:visited {
     padding: 20px;
 }
 
-.page-content {
-    height: fit-content;
-    padding-bottom: 20px;
-}
-
 /* Menu bar at home page */
 .menu-bar {
-    display: flex;
+    display: block;
     text-align: center;
-    margin-bottom: 40px;
 }
 
 .menu-bar a {
@@ -98,35 +73,25 @@ header a:visited {
     color: rgb(139, 139, 139);
 }
 
-/* HomePage (index) */
-.home-img-container {
-    height: 400px;
-    overflow: hidden;
+/* HomePage specific styles */
+.page-content-home {
+    padding: 10px 0;
+    overflow-y: scroll;
+    height: 50vh;
+    background-color: white;
 }
 
-.home-img-container img {
+.sambina-home-img {
+    display: flex;
     width: 100%;
+    height: 250px;
+    overflow: hidden;
+    align-items: center;
+    border-radius: 10px;
 }
 
-.home-grid p {
-    margin: 0;
-    padding: 30px;
-    line-height: 1.8rem;
-}
-
-.home-grid {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    background-color: #003564;
-    margin-bottom: 40px;
-}
-
-@media (max-width: 600px) {
-    .home-grid {
-        display: flex;
-        flex-direction: column;
-    }
-
+.sambina-home-img img {
+    max-width: 100%;
 }
 
 /* All pages */
@@ -149,7 +114,6 @@ header a:visited {
     height: 1.23rem;
     display: inline;
     min-width: fit-content;
-    color: #1d4669;
 }
 
 .country-tags-area {
@@ -176,7 +140,7 @@ header a:visited {
     margin-bottom: 20px;
     border-left: 10px solid rgb(177, 255, 247);
     padding: 17px;
-    background-color: #1d4669;
+    background-color: white;
     border-radius: 7px;
 }
 
@@ -209,65 +173,4 @@ header a:visited {
 
 .job-item-img {
     width: 70px;
-}
-
-/* Hobbies */
-.hobby {
-    display: flex;
-    width: 100%;
-    overflow-x: scroll;
-    margin-bottom: 20px;
-    /* Hide scrollbar for IE, Edge and Firefox */
-    -ms-overflow-style: none; /* IE and Edge */
-    scrollbar-width: none; /* Firefox */
-}
-
-.hobby::-webkit-scrollbar {
-    /* Hide scrollbar for Chrome, Safari and Opera */
-    display: none;
-}
-
-
-.hobby img {
-    height: 450px;
-    margin-right: 20px;
-}
-
-/* Timeline */
-.post {
-    background-color: #003564;
-    padding: 20px;
-    margin-bottom: 16px;
-    box-shadow: 0 3px 10px rgb(0 0 0 / 0.2);
-}
-
-form {
-    background-color: #003564;
-    padding: 20px;
-    box-shadow: 0 3px 10px rgb(0 0 0 / 0.2);
-}
-
-form input {
-    margin: 10px 0;
-    height: 2.5rem;
-    width: 100%;
-}
-
-#wrong-email-span {
-    background-color: rgba(255, 255, 0, 0.25);
-    font-style: italic;
-    border-radius: 5px;
-    padding: 2px 10px;
-}
-
-#timeline-submit-btn {
-    color: white;
-    height: 2.3rem;
-    background-color: #00203f;
-    border: 0;
-    margin-bottom: 0;
-}
-
-#timeline-submit-btn:hover {
-    cursor: pointer;
 }
