+++ app/templates/base.html
@@ -17,42 +17,25 @@
     />
 
     <link lang="sass" rel="stylesheet" href="./static/styles/main.css"/>
-    <link rel="icon"
-          href="data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22 viewBox=%220 0 100 100%22><text y=%22.9em%22 font-size=%2290%22>🦅</text></svg>">
+    <link rel="icon" href="data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22 viewBox=%220 0 100 100%22><text y=%22.9em%22 font-size=%2290%22>🦅</text></svg>">
 
     <link
             rel="stylesheet"
             href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.3/font/bootstrap-icons.css"
     />
-
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/normalize/8.0.1/normalize.css"
-          integrity="sha512-oHDEc8Xed4hiW6CxD7qjbnI+B07vDdX7hEPTvn9pSZO1bcRqHp8mj9pyr+8RVC2GmtEfI2Bi9Ke9Ass0as+zpg=="
-          crossorigin="anonymous" referrerpolicy="no-referrer"/>
     <title>
-        {% block title %}{% endblock %}
+        {% block title %}
+        {% endblock %}
     </title>
 </head>
 
 <body>
 <header class="nav-bar">
-    <a href="/" class="site-title-header">Oscar Miranda Escalante</a>
-    <div class="icons">
-        <a href="https://www.linkedin.com/in/oscarmires/" target="_blank" >
-            <i class="bi bi-linkedin"></i>
-        </a>
-        <a href="https://twitter.com/oscr_me" target="_blank">
-            <i class="bi bi-twitter"></i>
-        </a>
-        <a href="https://github.com/oscarmires" target="_blank">
-            <i class="bi bi-github"></i>
-        </a>
-    </div>
+    {% block header %}
+    {% endblock %}
 </header>
 
 <div class="page-container">
-    <h1 style="margin-top: 10px">
-        {{ title }}
-    </h1>
     <div class="page-content">
         {% block content %}
