--- app/templates/timeline.html
@@ -1,82 +1,0 @@
-{% extends 'base.html' %}
-{% import 'components.html' as components %}
-
-
-{% block title %}{{ title }}{% endblock %}
-
-{% block content %}
-    {{ components.menuBar(includeHome=true) }}
-    <form id="form">
-        <label for="name" class="timeline-form-label">Name:</label><br>
-        <input type="text" id="name" name="name" placeholder="Type your name" required><br>
-        <label for="email" class="timeline-form-label">Email:</label>
-        <span id="wrong-email-span" style="display: none"> ⚠️ Invalid email</span><br>
-        <input type="text" id="email" name="email" placeholder="Type your email address" required><br>
-        <label for="content" class="timeline-form-label">Content:</label><br>
-        <input type="text" id="content" name="content" required><br>
-        <input type="submit" value="Submit" id="timeline-submit-btn">
-    </form>
-
-    {% for post in posts %}
-        <div class="post">
-            <h2>{{ post["created_at"] }}</h2>
-            <i>By {{ post["name"] }} ({{ post["email"] }})</i>
-            <p>{{ post["content"] }}</p>
-        </div>
-    {% endfor %}
-
-    <script>
-        const url = '/api/timeline_post'
-        const form = document.getElementById('form');
-        form.addEventListener('submit', function (e) {
-            // Prevent default behavior:
-            e.preventDefault();
-
-            // Create payload as new FormData object:
-            const payload = new FormData(form);
-            console.log(payload.values())
-
-            // Post the payload using Fetch:
-            fetch('/api/timeline_post', {
-                method: 'POST',
-                body: payload,
-            })
-                .then(res => {
-                    if (!res.ok) {
-                        switch (res.status) {
-                            case 400:
-                                return res.text().then(text => {
-                                    throw new Error(text)
-                                });
-                            case 503:
-                                throw new Error("Exceeded time limits. Try again later.");
-                            default:
-                                return res.text().then(text => {
-                                    throw new Error(text)
-                                });
-                        }
-                    } else {
-                        return res.json()
-                    }
-                })
-                .then(data => {
-                    console.log(data)
-                    window.location.reload()
-                })
-                .catch(error => {
-                    console.log(error)
-                    if (error.message.includes("Invalid email")) {
-                        document.getElementById("wrong-email-span").style.display = "inline";
-                        const emailInput = document.getElementById("email")
-                        emailInput.style.backgroundColor = "#ffffd3"
-                        emailInput.focus()
-                    } else {
-                        alert(error)
-                    }
-                })
-        })
-
-
-    </script>
-
-{% endblock %}
