+++ app/templates/work_experience.html
@@ -1,32 +1,34 @@
-{% extends 'base.html' %}
 {% import 'components.html' as components %}
 
+{% extends 'base.html' %}
 
 {% block header %}
-    <span>Sambina Islam Aninta</span>
+{{ components.backHomeButton() }}
+<span>Sambina Islam Aninta</span>
 {% endblock %}
 
 <!---->
 {% macro job_item(job) -%}
-    <div class="job-item">
-        <div class="job-item-info">
-            <h2>{{ job.title }}</h2>
-            <span class="job-organization">{{ job.organization }}</span>
-            <span class="job-date">{{ job.date }}</span>
-            <pre>{{ job.description }}</pre>
-        </div>
-        <div class="job-item-logo">
-            <img class="job-item-img" src="{{ job.logo_url }}"/>
-        </div>
+<div class="job-item">
+    <div class="job-item-info">
+        <h2>{{ job.title }}</h2>
+        <span class="job-organization">{{ job.organization }}</span>
+        <span class="job-date">{{ job.date }}</span>
+        <pre>{{ job.description }}</pre>
     </div>
+    <div class="job-item-logo">
+        <img class="job-item-img" src="{{ job.logo_url }}"/>
+    </div>
+</div>
 {%- endmacro %}
 
-
-{% block title %}{{ title }}{% endblock %}
+{% block title %}
+{{ Title }}
+{% endblock %}
 
 {% block content %}
 
-    {{ components.menuBar(includeHome=true) }}
-    {% for job in job_list %} {{ job_item(job) }} {% endfor %}
+<h1>{{ title }}</h1>
+{% for job in job_list %} {{ job_item(job) }} {% endfor %}
 
 {% endblock %}
