--- docker-compose.prod.yml
@@ -1,38 +1,0 @@
-version: "3"
-services:
-  myportfolio:
-    container_name: myportfolio
-    build: .
-    restart: always
-    env_file:
-      - .env
-    depends_on:
-      - mysql
-
-  mysql:
-    container_name: mysql
-    image: mariadb
-    restart: always
-    env_file:
-      - .env
-    volumes:
-      - mydatabase:/var/lib/mysql
-
-  nginx:
-    container_name: nginx
-    image: jonasal/nginx-certbot
-    restart: always
-    environment:
-      - CERTBOT_EMAIL=oscar-me@outlook.com
-    ports:
-      - 80:80
-      - 443:443
-    volumes:
-      - nginx_secrets:/etc/letsencrypt
-      - ./user_conf.d:/etc/nginx/user_conf.d
-    depends_on:
-      - myportfolio
-
-volumes:
-  mydatabase:
-  nginx_secrets:
