--- docker-compose.yml
@@ -1,28 +1,0 @@
-version: "3"
-services:
-  myportfolio:
-    container_name: myportfolio
-    build: .
-    restart: always
-    env_file:
-      - .env
-    environment:
-      - FLASK_ENV=development
-    ports:
-      - "3000:5000"
-    volumes:
-      - .:/myportfolio
-    depends_on:
-      - mysql
-
-  mysql:
-    container_name: mysql
-    image: mariadb
-    restart: always
-    env_file:
-      - .env
-    volumes:
-      - mydatabase:/var/lib/mysql
-
-volumes:
-  mydatabase:
