+++ example.env
@@ -1,5 +1,1 @@
 URL=localhost:5000
-MYSQL_HOST=localhost
-MYSQL_USER=myportfolio
-MYSQL_PASSWORD=mypassword
-MYSQL_DATABASE=myportfoliodb
