--- run_test.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-$PWD/python3-virtualenv/bin/python -m unittest discover -v tests/
