--- user_conf.d/myportfolio.conf
@@ -1,51 +1,0 @@
-server {
-    listen 80;
-    server_name oscar-me.duckdns.org;
-
-    if ($host = oscar-me.duckdns.org) {
-        return 301 https://$host$request_uri;
-    }
-}
-
-# Rate limiting
-#   5r/s -> requests must wait 200ms
-limit_req_zone $binary_remote_addr zone=general_limit:10m rate=5r/s;
-#   6r/m -> requests must wait 10 seconds
-limit_req_zone $binary_remote_addr zone=timeline_post_limit:10m rate=6r/m;
-
-server {
-    listen 443 ssl;
-    server_name oscar-me.duckdns.org;
-
-    error_page 588 = @postrequests;
-
-    location @postrequests {
-        # Any POST request will be limited by timeline_post_limit
-        limit_req zone=timeline_post_limit;
-
-        proxy_pass http://myportfolio:5000;
-    }
-
-    location / {
-        # Handling burst
-        #   - If there's space in the queue, up to 20 requests may exceed the rate limit
-        #   - It's necessary since loading a single page generates ~15 requests (images, icons, styles...)
-        #   - "nodelay" parameter tells nginx that queued requests don't have to wait the min. time (30 secs.)
-        limit_req zone=general_limit burst=20 nodelay;
-
-        proxy_pass http://myportfolio:5000/;
-    }
-
-    location /api/timeline_post {
-        if ($request_method = POST) {
-            return 588;
-        }
-
-        proxy_pass http://myportfolio:5000/api/timeline_post;
-    }
-
-    # Load the certificate files.
-    ssl_certificate /etc/letsencrypt/live/myportfolio/fullchain.pem;
-    ssl_certificate_key /etc/letsencrypt/live/myportfolio/privkey.pem;
-    ssl_trusted_certificate /etc/letsencrypt/live/myportfolio/chain.pem;
-}
